hggtk/email: somewhat proper behavior
The dialog is hard-coded to always supply --desc so that
the patchbomb extension isn't forced to launch an editor,
which will only confuse the user.

It still sucks that the patch series description is not
always necessary and there's no way for the dialog to know
when it is superfluous.