hggtk/history: performance improvements
Re-use text buffer and file list store, only create tags once
Borrow hgview's method of diff-collection; all-at-once
A side-effect of this is that we now have diff-stats that we could
plot as an option.

History browser is now _as fast_ as hgview in large repository.  Faster
at startup since we limit the amount of history we read.