about: Make logo transparent
The old 92x50 png was taken from the homepage, and has a white
background.

The new png has transparency, which gives a nicer look for the
application logo in the about dialog. In addition to this, the
about dialog now has a separate 32x32 icon, which looks nicer
than a down-scaled version of the 92x50 image.