thgconfig: Switch to in-place editing model for "paths" page
The old editing model showed a list of aliases where you could
select an alias. You would then have to edit the alias in a
field in the bottom of the page.

In addition to this you would have to actively "Refresh" your
change to the list, in addition to committing it with the
"Apply" button.

The new editing model has only one level of commit: the "Apply"
button.

The patch also simplifies the internal data structures.
By relying more on gtk.ListStore and gtk.TreeView some
synchronization code could be removed.