commit: skip record process for "clean" commits
If the user has not rejected any of the change hunks for the file, do not
subject it to the backup-revert-patch-commit-recover cycle that record
normally does for all modified files, simply commit it.

We do this in TortoiseHg because it decreases the chance of the cycle
failing when a soure file has been locked by another application, a
Windows specific problem.