nautilus: more simplification work for menu generators
menuThg.get_commands() should not have to devine a cwd or repo
from the passed in file list, each platform could have different
schemes.  So this command is now passed in repo and cwd, which
will be common.  Also, make a separate function for non-repo
menus.

On the nautilus side, fix it up so this logic happens in one
place.  Also, send canonized paths to hgproc to make things
more readable to the user (makes paths relative to cwd)