status: prevent traceback when rename diffs are change-selected
This prevents tracebacks, but the behavior is still suboptimal.  The
rename diff header is tied to both the rename source and target, making
it a tri-state entity when it can only display two states.  The diff header
can be "selected" or not when one of the files is unselected depending on
the order they were toggled.

Fixes #77  More work necessary before 0.8 to clean up the rest.