rename: fix directory rename behavior
When moving directories, hg calls ui.status() to give messages
for every file that is moved.  The presense of output made the
rename dialog think there were errors, and thus it exited harshly.
All it needed to do was ignore the status() messages by setting
the quiet flag on the passed in ui() instance.

Bug reported by Gerald Reichenbach