thgconfig: apply changes before editing file
Make sure we don't quit when there are unsaved changes in the
dialog.  It seems like it might be better to turn the dialog
question around and ask them if they would like to save changes,
rather than asking to ask without saving.  That way a yes would
save and then edit.  But... this may confuse our current users.