QueryDirstate.cpp: cache last result, use it if not older than 1s
The shell extension calls for each item (file or folder) all
overlay handlers until a handler returns true. The handlers
are called in the order: unchanged, modified, added.

If an item is modifed or added, we now just take the last cached
status (from the previous call to the unchanged handler) and use
that -- provided it's not older than one second.

We also cache status=0 results. This is especially helpful for
non-repo items, as for these all three handlers are called
by the shell.