status: replace reselect hack with less onerous hack
Turns out it was selecting each row in turn, causing the dialog
to load diffs for each file in turn.. at a rate of ~1 second each.

Gtk's documented workaround for this: signal_handler_(un)block()
results in attribute erorrs (doesn't exist).  So I set an early out
variable to tell the selection change handler to not load anything.