shellext: fix a subtle bug with added files
This bug manifested itself reliably on 64 bit Windows 7 RC build 7100
with 64 bit explorer. The effect was that files most of the time
didn't have any overlay icon. Having a single added file seemed
to cure it.

The reason why this bug was discovered is that Windows 7 seems
to call the overlay handlers in more randomized order, compared to
Windows XP, where the handlers are called in a very deterministic
order (mostly determined by the order in which they are registered
in the registry, as it seems).

(Besides that Windows 7 calls the handlers from multiple threads,
see 0a4df15f49de.)