setup: migrate to VC8 for building shell extensions
setup.py in thg-installer will build the shell extensions directly inside
the source repository, like it does the locales.  It will copy the results
into the files\ directory where the ISS script will look for it.  We also
have the option of checking in those files as a convenience to other people
who might not have VC8.