shellext: return empty verbs in CShellExt::GetCommandString (issue 442)
"Rename File" from TortoiseHg submenu triggered a simple file rename
instead of hgtk rename.

It seems the "rename" verb conflicted with the "rename" commmand of explorer.

Return empty verb strings in GetCommandString for now, as it seems they are
not needed anyway.

If the verbs should be needed in the future, the command names in menuDescList
should probably be changed or get a distinctive prefix.