shellext: show myTortoiseClass as prefix in debug output of HgQueryDirstate
Example debug view output:

[520] [THG] [C] HgQueryDirstate: stat.lstat("W:\thg-stable\hgtk") -> stat.isdir is 0
             ^
             (myTortoiseClass is 'C' in this example)

Current state of the overlay codebase should show only lines with "[C]", as that's the
handler which is called first by the shell extension's host (normally explorer.exe).

Other handlers ('M', 'A', '?') are called afterwards and use the cached status result
of the preceding call, which normally produces no debug out.

(as tested on Windows XP)