iss: make installing the shell extension optional
Introducing components. The shell component requires admin login.
Shows a warning box that the shell component will not be available
for installing, if the installer is run by a non-admin.

The installer runs perfectly fine under non-admin and just
installs everything but the shell extension. Like this,
users with non-admin accounts are able to install mercurial
and TortoiseHg (dialogs!) command line commands ('hg' and 'hgtk').
In this case, the installer doesn't even force a restart.
This should also be quite useful for people just wanting bare
'hg' as well.

OS restart is only done if the shell extension is installed
(or uninstalled). So, if a non-admin runs the installer, there
will be no restart anyway.

Note that the 'restartreplace' Flag is specified *not* to work
if the installer is run by a non-admin (per Inno Setup's docu
*and* msdn documentation -- see function MoveFileEx, which
is used by Inno Setup).

Note: The diff for registry.iss looks like a near complete
rewrite, but it is not. I had to add '; Components: shell'
to almost all lines.

Furthermore, expanded shutdown of thgtaskbar.exe for uninstall
case.