thgmq: put toolbar on top, using horizontal layout
This frees up the space that was used by the lengthy vertical
toolbar on the left (we have already done the same change for
the "load more" and "load all" buttons in the history dialog).

The buttons in the toolbar on top use the left/right idiom
for "backwards" / "forward", which is known by left-to-right
written languages like English.

The buttons are layed out like this:

   |<      <      >       >|
   [1]    [2]    [3]     [4]

using the following tooltip texts:

  [1] "Unapply all patches"
  [2] "Unapply last patch"
  [3] "Apply next patch"
  [4] "Apply all patches"

For the case of using the MQWidget in the history dialog (which
currently is the only one where it is used):

The right arrows nicely point towards the changelist pane, to the
right of the queue pane.

When a patch is applied, it effectively "goes to the changelist pane
on the right". When a patch is unapplied, it effectively is "taken
away from the changelist pane, putting it back to the queue pane on
the left".