commit: Move parent frame to bottom of dialog
Since we now have an "advanced" frame, this IMHO looks a bit nicer.

Having the diff above the parents frame also fits with the mental model
of "the change is based on the parent revision(s)". Having status info
in (or at least near) the status bar is a good fit as well.

Maybe the M,R,?,I,A,!,C checkboxes should be moved to above the parent
frame next.

[1] AttributeError: 'GCommit' object has no attribute 'parents_frame'