thgmq: use gtk.ToolItem.set_tooltip instead of set_tooltip_text
set_tooltip_text is only available in PyGTK 2.12 and above. Since
we are restricted to PyGTK 2.10, we are not allowed to use it
(fixes de8a545808ca).

As a consequence, we have to specify a gtk.Tooltips object in
set_tooltip.

Note that gtk.Tooltips() is deprecated in PyGTK 2.12
(gdialog.py, line 227)