status: catch IOError from repo.status()
If we get an error from our background thread running repo.status(),
store the error message and display it in the status message of the
progress bar.  No files or diffs should be visible.  If the user fixes
the IO issue and hits refresh, they should get a functioning dialog.

Fixes #554