status: inactivate text diff and hunk selection tabs for non-MAR files
Clean ('C'), missing ('!'), unknown ('?') and ignored ('I') files don't
have diffs.

Diffs are only applicable for modified ('M'), added ('A') and removed ('R')
files.

This grays out the child page area and the tab labels of the text diff
and the hunk selection tabs in the right pane, if a file with a status
not in 'MAR' is selected.

This emphasizes that these tabs are not applicable for non-MAR files.

Note that the inactivated tabs can still be selected. It just gives
a bit better visual feedback.