history: enable file patterns
Strings entered in the 'File Patterns' field can be prefixed with one of the
mercurial match types (glob, re, path, relglob, relpath, relre). If not prefixed,
the default will either glob or path depending on the presense of one or more of
the characters '[{*?'.

File history mode is enabled when there is only one pattern item, it is of type
path or defaults to path and it is not an existing directory.