status: introduce 'fixed' file type and use it for '!' files
There's hardly ever a reason to hide deleted ('!') files
(note that 'hg help status' calls them 'missing').

Having deleted ('!') files is just a plain user error that
they should fix in the first place.

Let's not provide users the opportunity to suppress them
("Oh. I inadvertantly unchecked the 'deleted' files. So
I couldn't see that I forgot to hg remove them!")

So we don't provide a checkbox for it, just provide
the explanation for the '!' status code in form of a label.