update: use radio group for options
A combobox containing the option -C is just way too dangerous.
One inadvertant tiny turn on the mousewheel and you have -C
in the combobox. Forget checking it, and a click on "Update"
later your local changes are gone forever!

This is a perfect case for classic radio buttons, so let's use
them here instead of a combobox.

We have enough room for using radio buttons. Like this, the
user can read the texts of all three choices at once and all
the time.

Tweaking the texts while we're at it.