update: remove "Interactive" radio button and expander
The "Interactive" radio button is unneeded, we only need a checkmark
button for -C. If -C is not checkmarked, then we _do_ use the interactive
method.

The expander is dangerous. Imagine a user selects -C and closes the expander.
Then he gets distracted (phone call). Later, he gets back to the update dialog
but he has forgotten that he checked -C. He doesn't see that he checked -C
because it is hidden under the closed expander! So he clicks "Update" and
his changes are gone. The -C checkmark button (with the "no backup" text) must
*always* be visible. Given its importance, let's place it directly below the
combobox.

Furthermore, putting the -C checkmark button *alone* into an expander
would be pointless (remember, the "Interactive" radio is removed).

Since we don't have an expander anymore, there's nothing left to store in
settings anymore. So we can remove that as well.