update: use add_button with gtk.RESPONSE_OK for "Update"
The button with the response id gtk.RESPONSE_OK is the 'default
response button' (on Windows, this is called the default button,
see also gtk.Dialog.set_default_response).

The default response button gets a special button frame, so
that the user recognizes it as the default button.

This is the button that is executed when return is hit after
dialog open.

The use case here is: select a changeset in history dialog, then
do "Update" from context menu -> update dialog opens with the
changeset selected as target. Hitting the <return> (or <enter>)
key then updates to that revision and closes the dialog.