history: disable strip and qimport on mq and mq ancestor changesets
Previously repositories with mq changesets applied could have them stripped
causing mq to go into a "bad" state that would require manual update of the mq
state file to reflect the modified layout of the repository. This fix makes it
impossible to strip or qimport an ancestor to an applied mq changeset.