shellext: InitMenuMaps: ignore template entry from menuDescList
This change is not fixing a specific bug (at least not as far as I
can see). It also doesn't fix the problem I found when building
the shell extension with mingw (Issue 624).

It just makes not much sense to go ask for the translation
of a menu entry having the empty string as its name.

There will be no key in the registry having the empty
string as its last component, so the registry lookup will
fail anyway (as it will ask for the UI text and help text
at the parent node of the translation keys in the registry).

Also essentially doing

   MenuDescMap[''] = ...

later on in that function (as before this change), doesn't
make much sense IMHO.

I see no code part that is expecting an entry under the empty
string in MenuDescMap anyway. And if it would, it would most
likely be a bug anyway.