history: add "Toolbar" checkmark entry in "View" menu
This allows to turn the main toolbar off, saving a lot of vertical
space.

Unfortunately I had to use gtklib.idle_add_single_call to correctly
set the initial state of the toolbar visibility in prepare_display.

This has the drawback, that when the dialog is opened, the toolbar
is always shown for a short period (even if the state should be
invisible right from the beginning).

I tried using set_no_show_all(True) like the other bars, but this
just always produced an empty toolbar.