Make forking configurable in config.py and disable it by default on Posix.
The new default value can be overruled with the config setting
tortoisehg.hgtkfork = 1.

Unix has shell commands for running processes in the background, so the
built-in forking support isn't necessary. Very few Unix gui commands without
MDI or tabs forks, so hgtk will be more predictable and tradition-compliant if
it also doesn't fork.