history: reword UI texts in "pending bundle on exit" dialog
The previous question "Pull these incoming changes, or discard?"
didn't fit that well with the available buttons "Accept" and "Reject".

The text in the dialog used the words "pull" and "discard", but the
button labels are "Accept" and "Reject".

Consistently using the terms "accept" and "reject" in the dialog
text *and* the button labels now.

I also replaced the word "Incoming" with "new" in the dialog title,
since the bundle can have been added by using "Add Bundle..." from the
"Synchronize" menu as well (not just by using the "Incoming" command).