history: fix logview row selection
Currently we use get_selected_rows, which among other things is problematic
when people move the cursor using ctrl + mouse click or arrow keys, and when
the user uses in-line searching. Either the selection is ignored, or, worse,
the selection is displayed inconsistently with the actually selected row
(since the selection lags behind the emit of the cursor-changed signal).

Since it's a single-selection treeview, we switch to using get_cursor instead.