status: add accelerators for filetree toggle, next/previous diff page
Added keyboard accelerator Ctrl-u to toggle file selection.  The accelerator
simulates clicking the check box in the filetree header: toggling it checks or
unchecks all the files in the tree.

Added keyboard accelerators Ctrl-p and Ctrl-Shift-p to move to the
next/previous page in the diff notebook.

Extracted the keyboard accelerator wiring code to a helper method.

closes #289