visdiff: use null ctx as ancestor for copies(), when not merging
This fixes rename detection for non-merge revisions.  It does not address the
fact that renames are not found for files in merge revisions that were not
modified. That will require a separate fix.  I'm leaning towards showing only
status(ctx, ctx.p1()) in the visual diff window for merges, removing the second
status column, and adding a button to launch 3-way directory diffs if the
selected diff tool supports it.