status: restore ability to toggle display of deleted files
This started as a simpler change.  All that needed to be done was change the
True to a False in the 'deleted' allchecks tuple.  However, once that was done
it made the 'fixed' variable obsolete so I removed it.  This also makes the
comment above allchecks valid again.  Should we ever decide to disable it
again, we should backout this entire changeset.

Closes #847