history: get back performance of logview
Since 'changes' column is introduced on default branch, logivew has
faced performance problem.  A repo.status() calling is direct cause
of it, but it's not a root cause; 'changes' column is always added
even if 'changes' column is hidden.

Because History dialog passes all columns to set_columns() method of
treeview, treemodel is requested all data for columns.  So we have to
pass only displaying columns.  This means that treeview doesn't have
full order of columns, only displaying columns in 'self.columns'
variable.