logview: do not add all columns on init
If it adds all columns in construct_treeview() on init, all data for
columns are requested even if the column is hidden.  To prevent requesting
of high-cost value ('changes' column), we have to add only columns to be
shown really.

Note that, by this change, you must call set_columns() function before
displaying this.  TreeView is no longer add columns in default.