setup.py: set copyright info in version resource of exe's
Note that py2exe will only be able to parse the version and create
a version resource in the exe's if it is a plain release (e.g. 0.9.3).

For, for example, nightly builds, py2exe warns with (example):

  warning: py2exe: Version Info will not be included:
    could not parse version number '0.9.2+39-622cd4651185+20100201'

and does not produce any version resource at all in the exe's.

Reason: exe (and dll) files on Windows can only have four groups
of digits in the binary version resource (e.g. 999.999.999.999).