setup.py: fix version handling on Windows exe file creation
The "binary" file versions for exe/dll files on Windows must have the form
W.X.Y.Z (with W,X,Y,Z in 0..65535), as can be seen by the warning we
currently get from py2exe when doing builds on Windows that are not from
a plain tag (i.e. with a '+' in the full version string):

  warning: py2exe: Version Info will not be included:
    could not parse version number '0.9.2+316-bb7de261c6b7+20100201'

Fixed by stripping the version string given to setup() to the part before
the '+' on Windows.

Windows exe/dll files can take an arbitrary string in the 'product version'
part of the version resource, which can be given to py2exe in the
'product_version' parameter. So we specify the full version string there.