shellext: insert wait in terminate
If the pipe worked ok (=icon server was running), then we wait a bit to give
the overlay icon server a bit time to shut down.

This works around a race in the InstallValidate action of the installer, where
InstallValidate proceeds immediately to check if processes are still using
files (terminate is called in a custom action right before InstallValidate).

Previously, terminate didn't wait, which occasionally lead to InstallValidate
still listing the icon server in the user install UI as needed to shut down
(files in use).