protect filenames in command lines
Filenames that start with '--' cause trouble when used as positional
arguments on the command line since they are seen as a command line
flag. So

  hg clone foo --foobar

fails and result in questions like this:

  http://stackoverflow.com/questions/2523080/

There is a standard way to indicate that the following command line
arguments should be interpreted literally, namely by giving an
argument of '--'. So

  hg clone -- foo --foobar

works as expected.