textview: support undoable 'replace' operation
When set_text() API of UndoableTextBuffer is called, it recognizes
'replace' operation as a combined operation of 'delete' and 'insert'.
To undo/redo 'replace' operation, I needed to introduce UndoableReplace
object which consists of UndoableInsert and UndoableDelete objects.

When 'insert'/'delete' operation is happened, UndoableReplace object
will be generated if the previous object is 'delete'/'insert' operation
and has same time stamp.