htmllistview: multi-selection is quite nice for picking chunks
The look and feel of chunk selection is coming into focus.  We will only
present one file at a time for selection, and the header chunk will not be
selectable (will not be in the tree view itself).  MultiSelection is used
to toggle the chunks the user wants to operate with.  There will be buttons
for "Revert", "Move to Patch", and "Hold from Commit".

The first two will have immediate effect, first verifying the file on disk
matches the loaded file, then doing an in-memory revert+patch, then writing
to disk.  The latter will mark the chunks as "held" and they will not be
committed (via a memctx commit).  I'm not sure yet how I'll show "held"
chunks in the normal diff view.