workbench: do not reload automagically
Previously, the model was unconditionally reloaded when it detected a modified
repository (includes dirstate changes!).

IMHO, this is way too intrusive. Reloading might be quite costly as we know.
And currently, it even flips the selected cset and scrolls to top (which can sure
be improved, but it's still quite intrusive).

Let's demote this to a simple message in the status bar for now, which recommends
to reload the repo model.

TODO: improve update use case (implies reload, due to changed parent cset and
working dir).