workbench: implement folder drag to open repos
Just drag a folder or a file on the workbench window and it opens the enclosing repo
in a new repo tab!

The drag is accepted only if the path is inside a repo.

Strangely, I had to strip the first character (a '/') from the path. I only tried on
Windows. Maybe someone else can test this (and please tweak if needed) on Linux.