repowidget: connect showMessage signals to a new showMessage() member function
to be wired-up higher in the widget hierarchy.

Open problem to be solved:

Make sure only the active repowidget can send its messages through to the workbench
statusbar.

And we probably need a mechanism to cache the last message and display it in the status
bar as soon as the respective repowidget becomes the active again.