reporegistry: try to fix updating of reporegistry tree view
If a repo was opened by using menu "File -> Open repository" the tree view
widget was not updated and the black triangle in front of the
"All Repositories" node even disappeared. Which is quite disappointing.

Fixed for now by calling reset() for the view (TODO: find a less drastic workaround)
and posting an expand request via single shot timer (direct call doesn't work).

If someone with more Qt fu knows how to fix this better, please fix it.