qtlib: take color definitions from thg-color
The color extension gets aggravated if you use effect names it does not know
about; so if you want TortoiseHg to use a color different from that defined by
the color extension, but the define in [thg-color].  It will override any
definition found in [color].

For example, the default log.changeset effect is yellow, which looks quite awful
in a GUI window with a white background.  So I specify:

[thg-color]
log.changeset = brown

We may want to ship some default thg-color defines in our site-wide ini file.