commit: begin development on a Qt commit tool
commit.py will house a CommitWidget, which holds most of the commit tool logic.
It is essentially a wrapper for a StatusWidget plus the commit extras.

It will also house a CommitDialog, a lightweight QDialog wrapper for a Commit
Widget, providing all the framework for preserving geometry, progress/status
bar, running commands, and handling events. Everyone who begged for 'ESC' to
close the commit tool will finally be happy.  The CommitDialog will have no
toolbar, and will have a typical dialog button bar on the bottom.

The CommitWidget is also intended to be embedded in the WorkBench app, so it
will present a generic SIGNAL/SLOT interface to its parent application.  The
WorkBench app may place the commit button wherever it likes.