history: do not try to encode/decode filenames for file history
File names may exist in a repository in mixed encodings and just forcing them
to local encoding is wrong. Particularly when the local encoding is manually
set to something different than the file system encoding. It is a tenant of
Mercurial to always see file names as byte strings, not as characters.