settings: introduce a settings tool
This dialog does nothing useful yet, and could destroy your ini file.  So be careful
with it.  Left out 'sync' page, that should be handled within the WorkBench somewhere.
Left out 'extensions' page as it needs improvement before 1.1 so it works on Windows.
After 1.1, I'll port the finished extensions page to this Qt version.