workbench: have one commit widget per disk-repo
not per repo tab.

Moving the revisions splitter and the stacked widget from
repowidget to workbench and have the repowidget's add their
revDetailsWidget's to that single stacked widget.

Multiple repo tabs (if they show the same on-disk repo) may
now use the same commit widget.

As a side effect, the position of the horizontal splitter is
now the same, no matter which repo tab is active.