guess: add progress/status bar
My attempts to pass the data via DataWrapper objects was met with
SIGSEGV at random times, and bizarre tracebacks from signal handler
functions about QObject instances not having 'data' members.

Not sure what to make of it, but this impl seems to be solid.  Bits
of this (ProgUi) should eventually be moved to common code.