annotate: fix "annotate parent"
"Annotate Parent" would simply go to the next row in the log instead
of to the real parent.

Simplified the code; instead of manually trying to chase renames from
the top of the log (which wouldn't work across multiple renames), we
simply use the parent file context, which takes care of renames for us.

For now, we always follow the first parent only; will be improved in
a later  commit.