logcolumns: introduce a configuration dialog for log columns
Tries to be future-proof.  The list of 'all' columns is passed to the
dialog.  If columns are added or removed, it just works.  Columns are
enabled by checking them, re-ordered by dragging them (Qt makes this
stupid easy).

The dialog is connected to 'thg test', until someone connects it to the
workbench in an appropriate fashion.