repoview: completely nuke column resize triggered by resize events
All resize events for the repoview were causing it to resize the columns.  When
the user manually stretched a column, resizes were disabled for a short time.
(yuck).  If the user column stretch caused a horizontal scrollbar to appear,
this triggered a resize event and a call to resizeColumns() (yuckier).

Now, the columns are only resized explicitly when the model is filled.  Perhaps
we want to also make this feature available as a menu/accelerator action.