workbench: more robust column update after re-order
When you re-order or change columns, we do not need to reload the
graph but we do need to update the model and reset its cache.  This
avoids a race between reload() and resizeColumns().

This also required backing out a change Adrian just made, else the
last column is made the largest even when it should have a fixed size.