shelve: catch IOError from recovering backup files
On Windows, shutil.copyfile will raise an IOError if the file is opened by
another process.  Note that this is the second exception raised by the locked
file.  The first exception (during the shelve patching phase) caused it to
drop into this recovery code (copying backups back over the original files).

Refs #1303