chunks: be more careful during file size checks
In changeset db27310504c4, I made the status tool pass a wctx to check_max_diff
when it is operating on the working copy.  I did not realize a wctx will happily
generate a fctx for an unrevisioned file.  So when the commit tool tried to
query the diff status of a subrepository (directory), it make it all the way to
the point where it tried to read the file data, where it then encountered a
predictable IOError.

This changeset fixes the problems in two ways.  First, we explicitly verify that
we are inspecting a regular file.  Then we also catch EnvironmentErrors from the
fctx.data() call, since this could run into various file locks on Windows.

Closes #1340