shellext: invent "peek" state for overlays
If a file was edited, saved and then the edit undone and saved again, the
overlay for that file was stuck at "modified" until the user manually
triggered an "Update icons" or caused a "hg status".

Fixed by inventing a new "peek" state ('P') which causes an update of
.hg/dirstate via TortoiseHgOverlayServer. A file X is considered to be in
peek state, if the size of X and its size in the dirstate are the same but
the mtimes are different. This requires a look into the file contents to
tell if the file really has changed or not.

Note that Thgstatus::update() is now called with path (instead of cur.hgroot)
so that shell notifications are done for the affected file.