repomodel, repowidget: refactor graph load signal handlers
At refresh, the model loads 500 revisions then emits a filled signal.
Historically, this was trying to set the current revision (selecting the
changeset and showing a file).  But this is a bit dangerous.  The current
rev could have been > 500, causing a thread race.

Instead, recalc the columns at this early filled signal, and select the
row at the end when the model emits the loaded signal.

Startup should "feel" a lot snappier, even though loading takes about the
same time.  I hope to improve on this further.