repowidget: more tuning of graph load signals
Nothing really needs to happen when the graph is fully loaded any more.  If we
add a GUI element later to indicate load status, we can update it to full load.
For now, it does nothing.  Also remove arbitrary selection of row 0, which is
always the working directory row, which has a performance hit to select.