docklog: introduce a progress/logging dock widget
Mercurial commands which run inside the workbench should pipe their
ui.write(), ui.write_err(), and ui.progress() calls to this widget.

As a dock widget, it may be moved about or hidden by the user.

It is designed to support multiple progress bars at the same time, roughly as
a stack.