thread: pyqtSignal cleanup
I believe some of the random crashes I have seen were caused by the DataWrapper
class constructor, which passed None to QObject.__init__().  It seems we can
just derive the class from 'object' and not worry about side.

Part of this cleanup is sending the completion message as a tag-along to the
completion signal instead of in-band in the output stream.  This allows us to
filter it from rawoutput and prevents encoding bugs.