status: don't remember check status of files not shown
If you toggle display of 'Modified' files, we will forget how you
had them all checked.  This was the only way to "uncheck" files
that are no longer modified.

The filter entry does not have any effect on the check count, it's
only the refresh triggered by toggling MARID?! that can forget state.