repowidget: directly read parents from dirstate file
This is a fairly serious layering violation, from Mercurial's point of view,
but their API does not allow for invalidating only the cached working parents.

I do not want to invalidate the repo dirstate cache every time the file is
written; because I am afraid this will lead to much more I/O than is necessary.