thgrepo: watch MQ series rather than status
The MQ transactions which modify status will also modify changelog, so we only
need to monitor the latter, which is also the canonical version.

Sometimes MQ will change the series file and the changelog. This is always done
as a single locked transaction, so acquiring and releasing the wlock after
witnessing a changelog timestamp change will ensure we invalidate the repo to a
consistent state.