commit: ignore checked file list when qrefresh'ing
Since the commit tool is only aware of the working copy changed
files, providing a file list will remove files from the top patch
that are not currently dirty (nice, eh).

When we have decent MQ/patch tools, QRef support in the commit
tool may need to be ripped out.  It all depends on the direction
this goes.