status: use Qsci to show diffs and raw files
The QSci and lexer is much faster than parsing our diffs to HTML and
rendering them.  This also allows us to preview untracked files with syntax
highlighting.  And we pick up consistent tab widths and whitespace visibility.

However, fileview.py is a huge mess and we need to add subrepos to the
diff generator.  I believe fileview needs to be detached from the revmodel
and graph.

When this is all cleaned up, it should apply well to the manifest tool.