revset: use QsciScintilla as revset entry, begin to connect events
For the record, a QLineEdit worked very eratically for this purpose.  The
return value of le.getSelection() was variable depending on whether the widget
had focus or not (if it had focus, it never thought it had a selection).

The Qsci widget is a big win here.  It give brace matching, auto-completions,
and call tooltips.  I'll have to add back a button or two to this dialog.