sync, repowidget: initial implementation of incoming bundle preview
PROS: Bundles are now created with incoming changesets and the repowidget can
display those changesets.

CONS: The manifest and search widgets are unaware, refresh is not functional,
there is no accept/reject functionality, and context menus have yet to be
restricted. Oh, and there's no visual indication of which changesets are
incoming.