repowidget: fix memory leak when invoking context menu
A new QMenu object was being created every time a user right-clicked on the graph.
These are added automatically to the repo's children collection and so would not
be eligible for garbage collection.

To fix this problem, I only create a single context menu for the repo widget and
cache it for reuse on subsequent occasions.