qsci: turn off horizontalScrollBars by enabling wrap
Apparently Qsci maintains a good amount of its speed by having a fixed length
display buffer (based on my tests). Even when wrap is disabled, your text lines
will be truncated at some length. and the horizontalScrollBars cover this
width. By turning on character wrap, we get rid of the meaningless scrollbars
and display all text.

This is probably not the last word on this topic (we likely need to enable some
of the wrap visualization features and/or make this toggle configurable).