commit: use cur-message.txt again
Using last-message.txt turned out to be counterproductive.  Mercurial wrote
to the file during imports and other operations and it is rarely empty.  So I'm
switching back to our own file again for storing the message in-progress.  But
the commit tool reads cur-message.txt and includes it in the message history so
the user may use the message found there.