cmdui: launch console hg commands with --cwd
this way the user can enter file paths that are repo canonical, and
not absolute.

thg does not support --cwd, else we may want to use it there as well,
but users do not need to provide filename arguments to thg commands
very often, so the benefit would be mimimal.