visdiff: make history snapshot files read-only
This requires atexit cleanup code to make the files writeable again so they
can be deleted on all platforms (*cough* Windows *cough*)

Read-only is a nice attribute in that it indicates to editors and visual
diff tools that the file contents are historical and immutable.  In other words,
editing those files will have no useful effects.