cmenu: don't set a cwd for starting the overlay server
If the overlay server is not running on "update icons" from the cmenu,
the cmenu starts TortoiseHgOverlayServer.exe with the cwd set to the
directory of the cmenu action, e.g. C:\repoA.

This prevents the directory C:\repoA from being moved because it is
the cwd of TortoiseHgOverlayServer.exe (for as long as it runs).

I fixed this by setting the working dir to the TortoiseHg program
directory.