chunks: introduce a patch chunk browser
This widget is incomplete, it does not yet parse patches into selectable chunks
and it does not make the chunks selectable.  A pair of these widgets will be the
core of the new patch editor tool.  You'll be able to move chunks from one side
to the other or revert them.