repowidget: do not connect commitComplete to self.reload anymore
Because the repowidget is connected to repositoryChanged signals, it gets
notified of commits as soon as they happen.  This extra signal was leading to
a double refresh and the second refresh was set to reload the graph to -1 which
caused the entire changelog graph to be loaded and for columns to shift around.